requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Ollama Integration
ollama==0.3.3

# Async & Concurrency
asyncio-mqtt==0.16.2
aiohttp==3.10.5
aiocache==0.12.2

# CLI & Formatting
rich==13.8.1
click==8.1.7
typer==0.12.5

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0

# Utilities
python-multipart==0.0.12

numpy

# Voice Transcription
faster-whisper==1.0.3

# Oracle Database & Vector Store
oracledb==2.4.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
markdown==3.7

# Embeddings & RAG
sentence-transformers==3.1.1
transformers==4.45.2
torch==2.5.1

# Web Search & Browsing
duckduckgo-search==6.3.0
beautifulsoup4==4.12.3
trafilatura==1.12.2
lxml==5.3.0
requests==2.32.3
